Remove UnavailableActionDialogProps
After discussing the issue
(https://github.com/comit-network/comit-i/issues/49), we have come to
the conclusion that we always want to display the action payload.

Actions that can be unavailable now display the time to wait until
they will be available. Additionally, the MetaMask `SendTransaction`
button is disabled an Ethereum action is unavailable, since we don't
want to directly enable broadcasting of invalid transactions.