Split ToForm component into two TextField components
The two textfields are used in different contexts and thus a more
composable solution was required. Specifically, on the MakeLink
page, we want to show a tooltip if the peerID was automatically
filled into the textfield. With the ToForm component, this would not
be possible without making the whole form aware of this functionality,
even though it is only used in one of the call sites: The MakeLink page.